reads 2 temps, 3PTs & writes to 2 Inverters
Successfully reads the sensors and can switch the inverters on based on temp difference to an coded-in pool temp (23C).  Needs pressure on the system pressure transducer otherwise everything switched off.  Has diagnostics on the Serial Monitor showing data in the results array and a dot for each time it has completed a run through the modbus manager (9 reads, and 2 writes).